migrate(core-ca): implements full blown ca

+++ src/main/java/com/sandari/rain/application/usecases/ApplicationDeleteUser.java
@@ -1,7 +1,9 @@
 package com.sandari.rain.application.usecases;
 
+import java.util.Optional;
 import com.sandari.rain.application.components.ApplicationException;
 import com.sandari.rain.application.interfaces.IApplicationUserRepository;
+import com.sandari.rain.domain.interfaces.IDomainUser;
 
 public class ApplicationDeleteUser {
     private final IApplicationUserRepository userRepository;
@@ -11,6 +13,12 @@ public class ApplicationDeleteUser {
     }
 
     public void execute(Long id) throws ApplicationException {
+        Optional<IDomainUser> fetchedUser = this.userRepository.getUserById(id);
+
+        if(fetchedUser == null) {
+            throw new ApplicationException("User not exist", 404, true);
+        }
+
         this.userRepository.deleteUser(id);
     }
 }
